Code Cleanup (Java conventions and stuff)
