+++ build.gradle
@@ -5,7 +5,7 @@ sourceCompatibility = '1.8'
 [compileJava, compileTestJava]*.options*.encoding = 'UTF-8'
 
 if (!hasProperty('mainClass')) {
-    ext.mainClass = 'sortvisualiser.MainApp'
+    ext.mainClass = 'MainApp'
     mainClassName = ext.mainClass
 }
