+++ src/main/java/me/hopson/sortvisualiser/MidiSoundPlayer.java
@@ -1,45 +1,40 @@
-package sortvisualiser;
+package me.hopson.sortvisualiser;
 
+import javax.sound.midi.*;
 import java.util.ArrayList;
-import javax.sound.midi.Instrument;
-import javax.sound.midi.MidiChannel;
-import javax.sound.midi.MidiSystem;
-import javax.sound.midi.MidiUnavailableException;
-import javax.sound.midi.Synthesizer;
 
 /**
- * 
  * @author Matthew Hopson
  */
 public class MidiSoundPlayer {
+    private static int cachedIndex = -1;
     private final ArrayList<Integer> keys;
-    private Synthesizer synth;
     private final MidiChannel channel;
-    
-    private final int inputValueMaximum;
-    private static int CACHED_INDEX = -1;
 
-    public MidiSoundPlayer(int maxValue) {
+    private final int inputValueMaximum;
+    private Synthesizer synthesizer;
+
+    MidiSoundPlayer(int maxValue) {
         try {
-            synth = MidiSystem.getSynthesizer();
-            synth.open();
+            synthesizer = MidiSystem.getSynthesizer();
+            synthesizer.open();
         } catch (MidiUnavailableException ex) {
             ex.printStackTrace();
         }
         inputValueMaximum = maxValue;
-        
-        
+
+
         //Set up midi channel
-        channel = synth.getChannels()[0];
+        channel = synthesizer.getChannels()[0];
 
         //Electric grand piano sounds the best, so it tries to find that.
         //Sometimes it is not supported, so it defaults to 143
-        Instrument[] instruments = synth.getDefaultSoundbank().getInstruments();
-        if (CACHED_INDEX == - 1) {
+        Instrument[] instruments = synthesizer.getDefaultSoundbank().getInstruments();
+        if (cachedIndex == -1) {
             boolean found = false;
             int index;
-            for(index = 0; index < instruments.length; index++) {
-            	Instrument i = instruments[index];
+            for (index = 0; index < instruments.length; index++) {
+                Instrument i = instruments[index];
                 if (i.getName().equals("Electric Grand Piano")) {
                     found = true;
                     break;
@@ -48,11 +43,11 @@ public class MidiSoundPlayer {
             if (!found) {
                 index = 2;
             }
-            CACHED_INDEX = index;
+            cachedIndex = index;
         }
 
-        channel.programChange(instruments[CACHED_INDEX].getPatch().getProgram());
-        
+        channel.programChange(instruments[cachedIndex].getPatch().getProgram());
+
         //Set up keys
         keys = new ArrayList<>();
         //24 is the start of the somewhat audible major keys
@@ -67,14 +62,14 @@ public class MidiSoundPlayer {
             keys.add(i + 11);
         }
     }
-    
+
     private int convertToMajor(int v) {
-        float n = ((float)v / (float)inputValueMaximum);
-        int index = (int)(n * (float)keys.size());
+        float n = ((float) v / (float) inputValueMaximum);
+        int index = (int) (n * (float) keys.size());
         index = Math.max(1, Math.min(107, index));
         return keys.get(index);
     }
-    
+
     public void makeSound(int value) {
         int note = convertToMajor(value);
         channel.noteOn(note, 25);
