+++ src/main/java/me/hopson/sortvisualiser/screens/SortingVisualiserScreen.java
@@ -1,14 +1,12 @@
-package sortvisualiser.screens;
+package me.hopson.sortvisualiser.screens;
 
-import java.awt.BorderLayout;
+import me.hopson.sortvisualiser.MainApp;
+import me.hopson.sortvisualiser.SortArray;
+import me.hopson.sortvisualiser.algorithms.ISortAlgorithm;
+
+import javax.swing.*;
+import java.awt.*;
 import java.util.ArrayList;
-
-import javax.swing.JSpinner;
-import javax.swing.SpinnerNumberModel;
-import javax.swing.SwingWorker;
-import sortvisualiser.MainApp;
-import sortvisualiser.SortArray;
-import sortvisualiser.algorithms.ISortAlgorithm;
 
 /**
  * The main class for the sort visualiser GUI
@@ -21,24 +19,25 @@ public final class SortingVisualiserScreen extends Screen {
 
     /**
      * Creates the GUI
+     *
      * @param algorithms List of algorithms to run for visualisation
      * @param playSounds Whether or not you want the algorithm to play sounds
-     * @param app The main application
+     * @param app        The main application
      */
-    public SortingVisualiserScreen(ArrayList<ISortAlgorithm> algorithms, boolean playSounds, MainApp app) {
+    SortingVisualiserScreen(ArrayList<ISortAlgorithm> algorithms, boolean playSounds, MainApp app) {
         super(app);
         setLayout(new BorderLayout());
         sortArray = new SortArray(playSounds);
         add(sortArray, BorderLayout.CENTER);
         sortQueue = algorithms;
     }
-    
+
     private void longSleep() {
         try {
             Thread.sleep(1000);
         } catch (InterruptedException ex) {
             ex.printStackTrace();
-        } 
+        }
     }
 
     private void shuffleAndWait() {
@@ -46,24 +45,24 @@ public final class SortingVisualiserScreen extends Screen {
         sortArray.resetColours();
         longSleep();
     }
-    
+
     public void onOpen() {
         //This would block the EventDispatchThread, and so
         //it must run on a worker thread
-        SwingWorker<Void, Void> swingWorker = new SwingWorker<Void,Void>() {
+        SwingWorker<Void, Void> swingWorker = new SwingWorker<Void, Void>() {
             @Override
-            protected Void doInBackground() throws Exception {
+            protected Void doInBackground() {
                 try {
                     Thread.sleep(250);
                 } catch (InterruptedException ex) {
                     ex.printStackTrace();
-                } 
+                }
                 for (ISortAlgorithm algorithm : sortQueue) {
                     shuffleAndWait();
-                    
+
                     sortArray.setName(algorithm.getName());
                     sortArray.setAlgorithm(algorithm);
-        
+
                     algorithm.runSort(sortArray);
                     sortArray.resetColours();
                     sortArray.highlightArray();
@@ -72,13 +71,13 @@ public final class SortingVisualiserScreen extends Screen {
                 }
                 return null;
             }
-            
+
             @Override
             public void done() {
-                app.popScreen(); 
+                app.popScreen();
             }
         };
-        
+
         swingWorker.execute();
     }
 }
