--- src/main/java/sortvisualiser/screens/Screen.java
@@ -1,29 +1,0 @@
-package sortvisualiser.screens;
-
-import static sortvisualiser.MainApp.WIN_HEIGHT;
-import static sortvisualiser.MainApp.WIN_WIDTH;
-
-import java.awt.Dimension;
-
-import javax.swing.JPanel;
-
-import sortvisualiser.MainApp;
-
-/**
- * 
- * @author mhops
- */
-public abstract class Screen extends JPanel {
-    protected MainApp app;
-    
-    public Screen(MainApp app) {
-        this.app = app;
-    }
-    
-    @Override
-    public Dimension getPreferredSize() {
-        return new Dimension(WIN_WIDTH, WIN_HEIGHT);
-    }
-    
-    public abstract void onOpen();
-}
